feat: Implement graceful shutdown and fix clippy warnings
This change introduces a graceful shutdown mechanism for the GoatDB server. The server now listens for SIGINT and SIGTERM signals and will shut down gracefully when they are received.

A control script, scripts/db_control.sh, has been added to manage the server's lifecycle. The script can be used to start, stop, restart, and check the status of the server.

Additionally, this change addresses a large number of clippy warnings to improve code quality and ensure CI passes.

+++ src/goatkv/storage/wal_manager.rs
@@ -191,7 +191,7 @@ impl Iterator for WalIterator {
         }
 
         // 验证校验和
-        if WalManager::get_checksum(&key, key_len as u32, &value, value_len as u32) != checksum {
+        if WalManager::get_checksum(&key, key_len, &value, value_len) != checksum {
             return Some(Err(io::Error::new(
                 io::ErrorKind::InvalidData,
                 format!(
